// src/model/updateItemRequest.ts
export type UpdateItemRequest = {
  visible_id: string;
  name: string;
  product_number: string;
  description: string;
  purchase_year: number | null;
  purchase_price: number | null;
  durability: number | null;
  is_depreciation: boolean;
  connector: string;
  color: string;
  image: FileList;
};